Fix corruptions due to ItemStacks link to NBT save data: fix BUKKIT-3333
This is a vanilla issue - the 'tag' NBTTagComponent field of the ItemStack
assumes that it is OK to save a reference to an NBT supplied via load(),
and assumes it is OK to supply a reference to the internal field during
a save().  Neither is true, as Chunk NBT structures are required to be
read-only once created (due to being written asynchronously off the 
server thread AND due to the potential to be passed to a new Chunk if
the same chunk is reloaded before the writing of the NBT is completed by
the File I/O thread).  Keeping a live reference to the NBT copy passed in,
or to the NBT value passed back during saving, creates serious thread
safety issues which can result in corrupted data being written to the
world data files.

The specific issue here was surfaced by the recent change to use 
setName("") on the ItemStack.tag object - which, when a chunk was being
reloaded before its save was completed, resulted in name of the field
in the NBT being set to "" - which caused it to be saves as "" instead
of "tag", resulting in it not being properly reloaded in the future
(and the corresponding loss of all enchanges, names, or other metadata).