Ignore player spawn when The End for End portal use; Fixes BUKKIT-3517
CraftBukkit's implementation spawnWorld meta data for players causes the
spawnpoint command to be allowed to set player specific spawn points in
The End.  A player with a spawn point in The End will never be able to
leave The End.

This commit ignores the player specific spawn point for such a situation
allowing the code to use the default Normal world which then allows the
player to effectively exit The End through an End portal.

This commit also brings bed verification logic back in line with vanilla
when using an End portal to return to the Normal world.